Added tests for the new mutation; split auth tests into a separate file
