+++ src/admin/resolvers/mutations/ApprovedItem.integration.ts
@@ -15,6 +15,7 @@ import {
   IMPORT_APPROVED_ITEM,
   REJECT_APPROVED_ITEM,
   UPDATE_APPROVED_ITEM,
+  UPDATE_APPROVED_ITEM_AUTHORS,
   UPLOAD_APPROVED_ITEM_IMAGE,
 } from './sample-mutations.gql';
 import {
@@ -22,6 +23,7 @@ import {
   ApprovedItemAuthor,
   CreateApprovedItemInput,
   RejectApprovedItemInput,
+  UpdateApprovedItemAuthorsInput,
   UpdateApprovedItemInput,
 } from '../../../database/types';
 import { CuratedCorpusEventEmitter } from '../../../events/curatedCorpusEventEmitter';
@@ -461,99 +463,6 @@ describe('mutations: ApprovedItem', () => {
       ).to.equal(data?.updateApprovedCorpusItem.externalId);
     });
 
-    it('should succeed if user has access to one of scheduled surfaces', async () => {
-      // Set up auth headers with access to a single Scheduled Surface
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.NEW_TAB_CURATOR_ENGB}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers, eventEmitter);
-      await server.start();
-
-      // Set up event tracking
-      const eventTracker = sinon.fake();
-      eventEmitter.on(ReviewedCorpusItemEventType.UPDATE_ITEM, eventTracker);
-
-      const res = await server.executeOperation({
-        query: UPDATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      // Good to check for any errors before proceeding with the rest of the test
-      expect(res.errors).to.be.undefined;
-      const data = res.data;
-
-      // External ID should be unchanged
-      expect(data?.updateApprovedCorpusItem.externalId).to.equal(
-        item.externalId
-      );
-
-      // Updated properties should be... updated
-      expect(data?.updateApprovedCorpusItem).to.deep.include(input);
-
-      // The `updatedBy` field should now be the SSO username of the user
-      // who updated this record
-      expect(data?.updateApprovedCorpusItem.updatedBy).to.equal(
-        headers.username
-      );
-
-      // Check that the UPDATE_ITEM event was fired successfully:
-      // 1 - Event was fired once!
-      expect(eventTracker.callCount).to.equal(1);
-      // 2 - Event has the right type.
-      expect(await eventTracker.getCall(0).args[0].eventType).to.equal(
-        ReviewedCorpusItemEventType.UPDATE_ITEM
-      );
-      // 3- Event has the right entity passed to it.
-      expect(
-        await eventTracker.getCall(0).args[0].reviewedCorpusItem.externalId
-      ).to.equal(data?.updateApprovedCorpusItem.externalId);
-
-      await server.stop();
-    });
-
-    it('should fail if request headers are not supplied', async () => {
-      // With the default context, the headers are empty
-      const server = getServer(eventEmitter);
-      await server.start();
-
-      const result = await server.executeOperation({
-        query: UPDATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.data).to.be.null;
-      expect(result.errors).not.to.be.null;
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-
-    it("should fail if user doesn't have access to any of scheduled surfaces", async () => {
-      // Set up auth headers with access to something irrelevant here, such as collections
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.COLLECTION_CURATOR_FULL}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers, eventEmitter);
-      await server.start();
-
-      const result = await server.executeOperation({
-        query: UPDATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.data).to.be.null;
-      expect(result.errors).not.to.be.null;
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-
     it('should fail if sent an invalid topic', async () => {
       // this should be `HEALTH_FITNESS`
       input.topic = 'HEALTH FITNESS';
@@ -605,6 +514,73 @@ describe('mutations: ApprovedItem', () => {
       expect(result.errors?.[0].message).to.contain(
         'does not exist in "CorpusLanguage" enum.'
       );
+    });
+  });
+
+  describe('updateApprovedCorpusAuthorsItem mutation', () => {
+    let item: ApprovedItem;
+    let authors: ApprovedItemAuthor[];
+    let input: UpdateApprovedItemAuthorsInput;
+
+    beforeEach(async () => {
+      item = await createApprovedItemHelper(db, {
+        title: "3 Things Everyone Knows About LEGO That You Don't",
+        status: CuratedStatus.RECOMMENDATION,
+        language: 'EN',
+      });
+
+      authors = [
+        { name: 'Author One', sortOrder: 1 },
+        { name: 'Author Two', sortOrder: 2 },
+      ];
+
+      input = {
+        externalId: item.externalId,
+        authors,
+      };
+    });
+
+    it('should succeed if user has full access', async () => {
+      // Set up event tracking
+      const eventTracker = sinon.fake();
+      eventEmitter.on(ReviewedCorpusItemEventType.UPDATE_ITEM, eventTracker);
+
+      const res = await server.executeOperation({
+        query: UPDATE_APPROVED_ITEM_AUTHORS,
+        variables: { data: input },
+      });
+
+      // Good to check for any errors before proceeding with the rest of the test
+      expect(res.errors).not.to.exist;
+      const data = res.data;
+
+      // External ID should be unchanged
+      expect(data?.updateApprovedCorpusItemAuthors.externalId).to.equal(
+        item.externalId
+      );
+
+      // Updated properties should be... updated
+      expect(data?.updateApprovedCorpusItemAuthors.authors).to.deep.equal(
+        input.authors
+      );
+
+      // The `updatedBy` field should now be the SSO username of the user
+      // who updated this record
+      expect(data?.updateApprovedCorpusItemAuthors.updatedBy).to.equal(
+        headers.username
+      );
+
+      // Check that the UPDATE_ITEM event was fired successfully:
+      // 1 - Event was fired once!
+      expect(eventTracker.callCount).to.equal(1);
+      // 2 - Event has the right type.
+      expect(await eventTracker.getCall(0).args[0].eventType).to.equal(
+        ReviewedCorpusItemEventType.UPDATE_ITEM
+      );
+      // 3- Event has the right entity passed to it.
+      expect(
+        await eventTracker.getCall(0).args[0].reviewedCorpusItem.externalId
+      ).to.equal(data?.updateApprovedCorpusItemAuthors.externalId);
     });
   });
 
@@ -1087,313 +1063,6 @@ describe('mutations: ApprovedItem', () => {
       );
       // Check that the ADD_SCHEDULE event was not fired
       expect(addScheduleEventTracker.callCount).to.equal(0);
-    });
-  });
-});
-
-describe('mutations: ApprovedItem - authentication checks', () => {
-  const eventEmitter = new CuratedCorpusEventEmitter();
-
-  // a standard set of inputs for this mutation
-  const input: CreateApprovedItemInput = {
-    prospectId: '123-abc',
-    title: 'Find Out How I Cured My Docker In 2 Days',
-    url: 'https://test.com/docker',
-    excerpt: 'A short summary of what this story is about',
-    authors: [
-      { name: 'Mark Twain', sortOrder: 1 },
-      { name: 'Jane Austen', sortOrder: 2 },
-    ],
-    status: CuratedStatus.CORPUS,
-    imageUrl: 'https://test.com/image.png',
-    language: 'DE',
-    publisher: 'Convective Cloud',
-    topic: Topics.TECHNOLOGY,
-    source: CorpusItemSource.PROSPECT,
-    isCollection: false,
-    isTimeSensitive: true,
-    isSyndicated: false,
-  };
-
-  beforeAll(async () => {
-    await clearDb(db);
-  });
-
-  describe('createApprovedCorpusItem mutation', () => {
-    it('should succeed if user has access to one of scheduled surfaces', async () => {
-      // set up event tracking
-      const eventTracker = sinon.fake();
-      eventEmitter.on(ReviewedCorpusItemEventType.ADD_ITEM, eventTracker);
-
-      // Set up auth headers with access to a single Scheduled Surface
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.NEW_TAB_CURATOR_ENGB}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers, eventEmitter);
-      await server.start();
-
-      const result = await server.executeOperation({
-        query: CREATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.errors).to.be.undefined;
-      expect(result.data).not.to.be.null;
-
-      // Expect to see all the input data we supplied in the Approved Item
-      // returned by the mutation
-      expect(result.data?.createApprovedCorpusItem).to.deep.include(input);
-
-      // Check that the ADD_ITEM event was fired successfully:
-      // 1 - Event was fired once!
-      expect(eventTracker.callCount).to.equal(1);
-      // 2 - Event has the right type.
-      expect(await eventTracker.getCall(0).args[0].eventType).to.equal(
-        ReviewedCorpusItemEventType.ADD_ITEM
-      );
-      // 3- Event has the right entity passed to it.
-      expect(
-        await eventTracker.getCall(0).args[0].reviewedCorpusItem.externalId
-      ).to.equal(result.data?.createApprovedCorpusItem.externalId);
-
-      await server.stop();
-    });
-
-    it('should fail if request headers are not supplied', async () => {
-      // With the default context, the headers are empty
-      const server = getServer(eventEmitter);
-      await server.start();
-
-      const result = await server.executeOperation({
-        query: CREATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.data).to.be.null;
-      expect(result.errors).not.to.be.null;
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-
-    it("should fail if user doesn't have access to any of scheduled surfaces", async () => {
-      // Set up auth headers with access to something irrelevant here, such as collections
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.COLLECTION_CURATOR_FULL}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers, eventEmitter);
-      await server.start();
-
-      const result = await server.executeOperation({
-        query: CREATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.data).to.be.null;
-      expect(result.errors).not.to.be.null;
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-
-    it('should fail optional scheduling if user has no access to relevant scheduled surface', async () => {
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.NEW_TAB_CURATOR_DEDE}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers, eventEmitter);
-      await server.start();
-
-      // extra inputs for scheduling - note attempting to schedule onto the US New Tab
-      // while only having access to the German New Tab
-      input.scheduledDate = '2100-01-01';
-      input.scheduledSurfaceGuid = 'NEW_TAB_EN_US';
-
-      const result = await server.executeOperation({
-        query: CREATE_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.data).to.be.null;
-      expect(result.errors).not.to.be.null;
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-  });
-
-  describe('rejectApprovedItem mutation', () => {
-    it('should successfully reject an approved item when the user has access to at least one scheduled surface', async () => {
-      // Set up auth headers with access to a single Scheduled Surface
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.NEW_TAB_CURATOR_ENGB}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers);
-      await server.start();
-
-      const item = await createApprovedItemHelper(db, {
-        title: '15 Unheard Ways To Achieve Greater Terraform',
-        status: CuratedStatus.RECOMMENDATION,
-        language: 'EN',
-      });
-
-      const input: RejectApprovedItemInput = {
-        externalId: item.externalId,
-        reason: 'MISINFORMATION,OTHER',
-      };
-
-      const result = await server.executeOperation({
-        query: REJECT_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.errors).to.be.undefined;
-      expect(result.data).not.to.be.null;
-
-      // On success, mutation should return the deleted approved item.
-      // Let's verify the id.
-      expect(result.data?.rejectApprovedCorpusItem.externalId).to.equal(
-        item.externalId
-      );
-
-      await server.stop();
-    });
-
-    it('should throw an error when the user has no access any scheduled surface', async () => {
-      // Set up auth headers without access to any Scheduled Surface
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2`,
-      };
-
-      const server = getServerWithMockedHeaders(headers);
-      await server.start();
-
-      const input: RejectApprovedItemInput = {
-        externalId: 'test-id',
-        reason: 'MISINFORMATION,OTHER',
-      };
-
-      const result = await server.executeOperation({
-        query: REJECT_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.errors).not.to.be.undefined;
-      expect(result.data).to.be.null;
-
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-
-    it('should throw an error when the user has only read-only access', async () => {
-      // Set up auth headers with read-only access
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,${MozillaAccessGroup.READONLY}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers);
-      await server.start();
-
-      const input: RejectApprovedItemInput = {
-        externalId: 'test-id',
-        reason: 'MISINFORMATION,OTHER',
-      };
-
-      const result = await server.executeOperation({
-        query: REJECT_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.errors).not.to.be.undefined;
-      expect(result.data).to.be.null;
-
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-
-    it('should throw an error when the request headers are undefined', async () => {
-      // pass in empty object for headers
-      const server = getServerWithMockedHeaders({});
-      await server.start();
-
-      const input: RejectApprovedItemInput = {
-        externalId: 'test-id',
-        reason: 'MISINFORMATION,OTHER',
-      };
-
-      const result = await server.executeOperation({
-        query: REJECT_APPROVED_ITEM,
-        variables: { data: input },
-      });
-
-      expect(result.errors).not.to.be.undefined;
-      expect(result.data).to.be.null;
-
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
-    });
-  });
-
-  describe('uploadApprovedCuratedCorpusItemImage mutation', () => {
-    const testFilePath = __dirname + '/test-image.jpeg';
-
-    beforeEach(() => {
-      writeFileSync(testFilePath, 'I am an image');
-    });
-
-    afterEach(() => {
-      unlinkSync(testFilePath);
-    });
-
-    it('should not succeed if user does not have write to corpus privileges', async () => {
-      const headers = {
-        name: 'Test User',
-        username: 'test.user@test.com',
-        groups: `group1,group2,${MozillaAccessGroup.READONLY}`,
-      };
-
-      const server = getServerWithMockedHeaders(headers);
-      await server.start();
-
-      const image: Upload = new Upload();
-
-      image.resolve({
-        filename: 'test.jpg',
-        mimetype: 'image/jpeg',
-        encoding: '7bit',
-        createReadStream: () => createReadStream(testFilePath),
-      });
-
-      const result = await server.executeOperation({
-        query: UPLOAD_APPROVED_ITEM_IMAGE,
-        variables: {
-          image: image,
-        },
-      });
-
-      expect(result.data).to.be.null;
-      expect(result.errors).not.to.be.null;
-      expect(result.errors?.[0].message).to.contain(ACCESS_DENIED_ERROR);
-
-      await server.stop();
     });
   });
 });
