+++ src/admin/resolvers/mutations/sample-mutations.gql.ts
@@ -27,6 +27,17 @@ export const UPDATE_APPROVED_ITEM = gql`
   ${CuratedItemData}
 `;
 
+export const UPDATE_APPROVED_ITEM_AUTHORS = gql`
+  mutation updateApprovedCorpusItemAuthors(
+    $data: UpdateApprovedCorpusItemAuthorsInput!
+  ) {
+    updateApprovedCorpusItemAuthors(data: $data) {
+      ...CuratedItemData
+    }
+  }
+  ${CuratedItemData}
+`;
+
 export const REJECT_APPROVED_ITEM = gql`
   mutation rejectApprovedItem($data: RejectApprovedCorpusItemInput!) {
     rejectApprovedCorpusItem(data: $data) {
